Remove driver DLL dependency from generated code
This is a workaround for issue #20, which appears to be the opposite of #19. In the latter, the driver was given .net core 3.1 references when it was trying to target 3.0. This time, the driver is given 3.0 references while trying to target 3.1.
It looks like LinqPad is downloading the 3.1 version of the driver, which tries to generate a dynamic assembly by targeting 3.1 too. This is necessary because the generated code references the driver assembly, which in this scenario targets 3.1. However, LinqPad is providing reference assemblies for .net core 3.0, which results in a compiler error.
I talked about the problem with Joe Albahari and hope that in the future the driver will be able to request version-specific reference assemblies from LinqPad. In the meanwhile, since the dynamic assembly should be 3.0-compatible if it doesn't reference the 3.1 driver itself, I have removed the reference and am now compiling the required classes with the rest of the code. This way, whether LinqPad provides 3.0 or 3.1 references, the generated code is totally agnostic and works with both.